Program to check serial numbers' validity
This program's aim is to check serial numbers' validity. Firstly, it asks the user for a serial number with 11 digits. Then, it adds the first 10 digits by after multiplying the singular ones with 3 and plural ones with 2. If the last digit of the sum is equal to the last digit of the serial number, it is is valid. Otherwise, it is not.