Runoff election program.
It is similar to the previous plurality election program. But in this version, being plurality is not enough. People vote for their 2nd and 3rd preference too. So, even if a candidate has not gotten the plurality, they can still win. For example, A candidate has gotten 4 first preference votes, B candidate has gotten 5 first preference votes. But Every voter has voted for the A candidate as their first or second preference, which means everyone would be happy if the A candidate has won. But most of the voters have voted for the B candidate as their last preference, only 5 voted them as their first choice. In this case, only 5 people would be happy if the B candidate has won. Therefore, first preference votes are not the only important thing to look at in this election type.